design update and colocate model components
