+++ frontend/app/onboarding/_components/onboarding-card.tsx
@@ -18,13 +18,13 @@ import {
 } from "@/app/api/queries/useGetSettingsQuery";
 import { useGetTasksQuery } from "@/app/api/queries/useGetTasksQuery";
 import type { ProviderHealthResponse } from "@/app/api/queries/useProviderHealthQuery";
+import { useDoclingHealth } from "@/components/docling-health-banner";
 import {
   EMBEDDING_PROVIDER_ORDER,
   getProviderChrome,
   LLM_PROVIDER_ORDER,
   orderProviders,
-} from "@/app/settings/_helpers/model-helpers";
-import { useDoclingHealth } from "@/components/docling-health-banner";
+} from "@/components/models/model-helpers";
 import { Button } from "@/components/ui/button";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import {
@@ -599,7 +599,7 @@ const OnboardingCard = ({
                 value={modelProvider}
                 onValueChange={handleSetModelProvider}
               >
-                <TabsList className="mb-1 pb-3 w-full justify-start gap-1 overflow-x-auto">
+                <TabsList className="mb-1">
                   {tabProviders.map((providerKey) => {
                     const chrome = getProviderChrome(
                       providerKey,
@@ -616,7 +616,9 @@ const OnboardingCard = ({
                           error &&
                             selected &&
                             "data-[state=active]:border-destructive",
-                          "min-w-40",
+                          // Fixed 3-up basis so every card is the same width
+                          // (like a grid) while flex still fills the row.
+                          "min-w-52 grow-0 basis-[calc((100%_-_1.5rem)/3)]",
                         )}
                       >
                         <TabTrigger
