+++ frontend/components/icons/azure-openai-logo.tsx
@@ -2,9 +2,8 @@
  * Azure OpenAI Service, as Microsoft draws it (az-icons.com/icon/azure-openai).
  *
  * Distinct from `ai-foundry-logo`: Microsoft draws the two Azure model services
- * differently. The gradient id is fixed rather than generated: every instance
- * defines the same stops, so when a page renders several of these and the
- * browser resolves `url(#...)` to the first one, they all paint identically.
+ * differently. Painted with `currentColor`, like `anthropic-logo`/`openai-logo`,
+ * so it can go from grayscale to full color by changing the parent's text color.
  */
 export default function AzureOpenAILogo(props: React.SVGProps<SVGSVGElement>) {
   return (
@@ -12,28 +11,13 @@ export default function AzureOpenAILogo(props: React.SVGProps<SVGSVGElement>) {
       xmlns="http://www.w3.org/2000/svg"
       width="16"
       height="16"
-      viewBox="0 0 18 18"
-      fill="none"
+      viewBox="0 0 16 16"
       {...props}
     >
       <title>Azure OpenAI Logo</title>
-      <defs>
-        <radialGradient
-          id="azure-openai-mark"
-          cx="-67.981"
-          cy="793.199"
-          r=".45"
-          gradientTransform="translate(-17939.03 20368.029) rotate(45) scale(25.091 -34.149)"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop offset="0" stopColor="#83b9f9" />
-          <stop offset="1" stopColor="#0078d4" />
-        </radialGradient>
-      </defs>
       <path
-        d="m0,2.7v12.6c0,1.491,1.209,2.7,2.7,2.7h12.6c1.491,0,2.7-1.209,2.7-2.7V2.7c0-1.491-1.209-2.7-2.7-2.7H2.7C1.209,0,0,1.209,0,2.7ZM10.8,0v3.6c0,3.976,3.224,7.2,7.2,7.2h-3.6c-3.976,0-7.199,3.222-7.2,7.198v-3.598c0-3.976-3.224-7.2-7.2-7.2h3.6c3.976,0,7.2-3.224,7.2-7.2Z"
-        fill="url(#azure-openai-mark)"
-        strokeWidth="0"
+        d="M0 2.4V13.6C0 14.9253 1.07467 16 2.4 16H13.6C14.9253 16 16 14.9253 16 13.6V2.4C16 1.07467 14.9253 0 13.6 0H2.4C1.07467 0 0 1.07467 0 2.4ZM9.6 0V3.2C9.6 6.73422 12.4658 9.6 16 9.6H12.8C9.26578 9.6 6.40089 12.464 6.4 15.9982V12.8C6.4 9.26578 3.53422 6.4 0 6.4H3.2C6.73422 6.4 9.6 3.53422 9.6 0Z"
+        fill="currentColor"
       />
     </svg>
   );
