+++ frontend/components/models/model-select-item.tsx
